Move "Include" class to generator.common
